Rename file that is no longer just the node contributor
